requirements.txt
torch
torchvision
sentencepiece
streamlit
sacrebleu
decord
opencv-python
pyyaml
tqdm
numpy
pandas
scikit-learn
